yolo-trainer/requirements.txt
ultralytics==8.3.12
mlflow==2.14.1
boto3==1.34.125
PyYAML==6.0.1
opencv-python-headless==4.10.0.84
pycocotools==2.0.7
onnx==1.16.1
onnxsim==0.4.36
onnxruntime-gpu==1.18.0
tqdm==4.66.4
